Setup the express server, added hot reloading support and configured Postgres DB connection
Setup the express server along with establishing the connection with local Postgres DB. Specified the Postgres DB connection info in ormconfig.json. Also added hot reloading support in the package.json file for development (this will reload the server if changes are made and hence speed up the development activity).